Conflicts:
	runtime/ruleset.c

+++ runtime/stream.c
@@ -16,7 +16,7 @@
  * it turns out to be problematic. Then, we need to quasi-refcount the number of accesses
  * to the object.
  *
- * Copyright 2008, 2009 Rainer Gerhards and Adiscon GmbH.
+ * Copyright 2008-2012 Rainer Gerhards and Adiscon GmbH.
  *
  * This file is part of the rsyslog runtime library.
  *
@@ -361,7 +361,7 @@ static rsRetVal strmCloseFile(strm_t *pThis)
 		pThis->fdDir = -1;
 	}
 
-	if(pThis->bDeleteOnClose) {
+	if(pThis->bDeleteOnClose && pThis->pszCurrFName != NULL) {
 		if(unlink((char*) pThis->pszCurrFName) == -1) {
 			char errStr[1024];
 			int err = errno;
@@ -369,13 +369,11 @@ static rsRetVal strmCloseFile(strm_t *pThis)
 			DBGPRINTF("error %d unlinking '%s' - ignored: %s\n",
 				   errno, pThis->pszCurrFName, errStr);
 		}
-	}
-
-	pThis->iCurrOffs = 0;	/* we are back at begin of file */
-	if(pThis->pszCurrFName != NULL) {
 		free(pThis->pszCurrFName);	/* no longer needed in any case (just for open) */
 		pThis->pszCurrFName = NULL;
 	}
+
+	pThis->iCurrOffs = 0;	/* we are back at begin of file */
 
 	RETiRet;
 }
